get rid of CMSDEPRECATED_X warnings in SLHCUpgradeSimulations/Geometry
